fix: eliminate remaining compiler warnings
- C4459: Rename source_id/target_id params to src_id/tgt_id in dynamic_graph.hpp edge constructor to avoid shadowing CPO declarations
- C4389: Add static_cast for signed/unsigned comparisons in graph_cpo.hpp find_vertex_edge, test_dynamic_graph_cpo_vofl.cpp degree checks, and partition comparisons
- C4244: Change vector types in test_target_cpo.cpp and test_partition_id_cpo.cpp to use size_t for first element to match return types
- C4296: Remove always-true >= 0 check on unsigned type in test_adjacency_list_vertex_concepts.cpp

+++ tests/test_partition_id_cpo.cpp
@@ -289,7 +289,7 @@ TEST_CASE("partition_id(g,u) - integration with vertex_id", "[partition_id][cpo]
     auto verts = vertices(graph);
     
     // All vertices have different IDs but same partition
-    std::vector<std::pair<int, int>> id_partition_pairs;
+    std::vector<std::pair<size_t, int>> id_partition_pairs;
     for (auto v : verts) {
         auto vid = vertex_id(graph, v);
         auto pid = partition_id(graph, v);
@@ -299,9 +299,9 @@ TEST_CASE("partition_id(g,u) - integration with vertex_id", "[partition_id][cpo]
     REQUIRE(id_partition_pairs.size() == 3);
     
     // Different vertex IDs
-    REQUIRE(id_partition_pairs[0].first == 0);
-    REQUIRE(id_partition_pairs[1].first == 1);
-    REQUIRE(id_partition_pairs[2].first == 2);
+    REQUIRE(id_partition_pairs[0].first == size_t(0));
+    REQUIRE(id_partition_pairs[1].first == size_t(1));
+    REQUIRE(id_partition_pairs[2].first == size_t(2));
     
     // Same partition for all
     REQUIRE(id_partition_pairs[0].second == 0);
